+++ app/services/tnt_import/contact_import.rb
@@ -74,8 +74,7 @@ class TntImport::ContactImport
     contact.updated_at = parse_date(row['LastEdit']) if @override
     contact.created_at = parse_date(row['CreatedDate']) if @override
 
-    contact.add_to_notes(row['Notes'])
-    contact.add_to_notes("Children: #{row['Children']}") if must_add_children?(row, contact)
+    add_notes(contact, row)
 
     contact.pledge_amount = row['PledgeAmount'] if @override || contact.pledge_amount.blank?
     # PledgeFrequencyID: Since TNT 3.2, a negative number indicates a fequency in days. For example: -11 would be a frequency of 11 days. For now we are ignoring negatives.
@@ -143,6 +142,22 @@ class TntImport::ContactImport
     company.save!
     donor_account.update_attribute(:master_company_id, company.master_company_id) unless donor_account.master_company_id == company.master_company.id
     company
+  end
+
+  def add_notes(contact, row)
+    contact.add_to_notes(row['Notes'])
+    contact.add_to_notes("Children: #{row['Children']}") if must_add_children?(row, contact)
+
+    # These social fields don't have equivalents in MPDX, so we'll add them to notes:
+
+    contact.add_to_notes("Other Social: #{row['SocialWeb4']}") if row['SocialWeb4']
+    contact.add_to_notes("Spouse Other Social: #{row['SpouseSocialWeb4']}") if row['SpouseSocialWeb4']
+
+    contact.add_to_notes("Voice/Skype: #{row['VoiceSkype']}") if row['VoiceSkype']
+    contact.add_to_notes("Spouse Voice/Skype: #{row['SpouseVoiceSkype']}") if row['SpouseVoiceSkype']
+
+    contact.add_to_notes("IM Address: #{row['IMAddress']}") if row['IMAddress']
+    contact.add_to_notes("Spouse IM Address: #{row['SpouseIMAddress']}") if row['SpouseIMAddress']
   end
 
   def true?(val)